+++ nemo/collections/asr/modules/__init__.py
@@ -27,4 +27,3 @@ from nemo.collections.asr.modules.conv_asr import (
     SpeakerDecoder,
 )
 from nemo.collections.asr.modules.lstm_decoder import LSTMDecoder
-from nemo.collections.asr.modules.subsampling import ConvSubsampling
